[DAY7] Terminé, pas optimisé...
# À voir pour éventuellement utiliser un pool de list<size_t> pour éviter les allocs

+++ Src/Day4.cpp
@@ -15,7 +15,7 @@ Day4::Day4()
 	auto lIt{ mFoundImage.begin() };
 	for (;;)
 	{
-		if (std::distance(lIt, mFoundImage.end()) > mWidth)
+		if (static_cast<size_t>(std::distance(lIt, mFoundImage.end())) > mWidth)
 		{
 			lIt += mWidth;
 		}
